+++ openghg_inversions/basis/algorithms/_quadtree.py
@@ -149,9 +149,9 @@ def get_quadtree_basis(fps: np.ndarray, nbasis: int, seed: int | None = None) ->
     cost = 1e6
     pwr = 0
 
-    fps = fps/np.sum(fps) # normalize to improve optimization stability
+    fps = fps / np.sum(fps)  # normalize to improve optimization stability
 
-    search_max = 10 * np.sum(fps)
+    search_max = 10.0
     while cost > 3.0:
         optim = scipy.optimize.dual_annealing(
             qtoptim, np.expand_dims([0, search_max / 2**pwr], axis=0), seed=seed