+++ openghg_inversions/postprocessing/countries.py
@@ -66,9 +66,6 @@ paris_regions_dict = {
         "NW_EU": ["BEL", "DEU", "DNK", "FRA", "GBR", "IRL", "LUX", "NLD"],
         "NW_EU2": ["BEL", "DEU", "FRA", "GBR", "IRL", "LUX", "NLD"],
         "NW_EU_CONTINENT": ["BEL", "DEU", "FRA", "LUX", "NLD"],
-    },
-    "eastasia":{
-        "EASTERN_ASIA":["EChi1", "PRK", "KOR", "JPN"]
     }
     }